Move abstraction specs.
