+++ Qowaiv.Validation.sln
@@ -24,8 +24,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "props", "props", "{696E5B2C
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Qowaiv.Validation.Abstractions", "src\Qowaiv.Validation.Abstractions\Qowaiv.Validation.Abstractions.csproj", "{71FE5C6E-2A11-468E-9E8F-AC4B3A780A33}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Qowaiv.Validation.Abstractions.UnitTests", "test\Qowaiv.Validation.Abstractions.UnitTests\Qowaiv.Validation.Abstractions.UnitTests.csproj", "{6C74ECDC-D724-4794-905F-609A27A42F99}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Qowaiv.Validation.Fluent", "src\Qowaiv.Validation.Fluent\Qowaiv.Validation.Fluent.csproj", "{67DC037A-0782-4D36-9A5F-1116062E77E6}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Qowaiv.Validation.DataAnnotations", "src\Qowaiv.Validation.DataAnnotations\Qowaiv.Validation.DataAnnotations.csproj", "{9F6C3C67-299E-4C60-B8E9-14A7B6B59627}"
@@ -69,10 +67,6 @@ Global
 		{71FE5C6E-2A11-468E-9E8F-AC4B3A780A33}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{71FE5C6E-2A11-468E-9E8F-AC4B3A780A33}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{71FE5C6E-2A11-468E-9E8F-AC4B3A780A33}.Release|Any CPU.Build.0 = Release|Any CPU
-		{6C74ECDC-D724-4794-905F-609A27A42F99}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6C74ECDC-D724-4794-905F-609A27A42F99}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6C74ECDC-D724-4794-905F-609A27A42F99}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6C74ECDC-D724-4794-905F-609A27A42F99}.Release|Any CPU.Build.0 = Release|Any CPU
 		{67DC037A-0782-4D36-9A5F-1116062E77E6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{67DC037A-0782-4D36-9A5F-1116062E77E6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{67DC037A-0782-4D36-9A5F-1116062E77E6}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -107,7 +101,6 @@ Global
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{696E5B2C-1F1D-4ACD-BBD4-40594FC7BF11} = {A4C3C9FF-4EED-41FE-8849-C896429254B5}
-		{6C74ECDC-D724-4794-905F-609A27A42F99} = {40C1E942-8DB1-4E58-8FB4-71F5EDA66029}
 		{C036985A-F17D-405A-B3FD-37F30FA154FB} = {A4C3C9FF-4EED-41FE-8849-C896429254B5}
 		{8E1F869D-393B-4F6C-A3C0-1D1D336F9FA8} = {C036985A-F17D-405A-B3FD-37F30FA154FB}
 		{F74D9EF2-39B5-40EB-8C44-1D6015C38D70} = {A4C3C9FF-4EED-41FE-8849-C896429254B5}
