+++ specs/Qowaiv.Validation.Specs/Properties/GlobalUsings.cs
@@ -8,3 +8,4 @@ global using System.Collections;
 global using System.Collections.Generic;
 global using System.Globalization;
 global using System.Linq;
+global using System.Threading.Tasks;
