--- test/Qowaiv.Validation.Abstractions.UnitTests/Qowaiv.Validation.Abstractions.UnitTests.csproj
@@ -1,29 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <Import Project="..\..\props\test.props" />
-  
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-  </PropertyGroup>
-  
-  <ItemGroup>
-    <ProjectReference Include="..\..\src\Qowaiv.Validation.Abstractions\Qowaiv.Validation.Abstractions.csproj" />
-    <ProjectReference Include="..\..\src\Qowaiv.Validation.TestTools\Qowaiv.Validation.TestTools.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="FluentAssertions" Version="6.6.0" />
-    <PackageReference Include="FluentAssertions.Analyzers" Version="0.17.2">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-    </PackageReference>
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.1.0" />
-    <PackageReference Include="nunit" Version="3.13.3" />
-    <PackageReference Include="NUnit3TestAdapter" Version="4.2.1">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-    </PackageReference>
-    <PackageReference Include="Qowaiv.TestTools" Version="6.0.0" />
-  </ItemGroup>
-
-</Project>
